+++ src/actions/careerPageActions.js
@@ -1,0 +1,10 @@
+const updateTabIndex = (value) => {
+  return {
+    type: 'UPDATE_TAB_INDEX',
+    payload: value
+  }
+}
+
+export default {
+  updateTabIndex
+}