+++ src/components/Home/CareerSection/CareerSection.js
@@ -1,13 +1,9 @@
 import { Container, Grid, Typography } from '@material-ui/core'
-import withWidth from '@material-ui/core/withWidth'
 import PropTypes from 'prop-types'
 import React from 'react'
-import { ProjectCard } from '../ProjectCard/ProjectCard'
-import { useStyles } from './style'
+import { ProjectSummaryCard } from '../ProjectSummaryCard/ProjectSummaryCard'
 
-const CareerSection = ({ width, projects }) => {
-  const classes = useStyles()
-
+export const CareerSection = ({ projects }) => {
   return (
     <Container>
       <Typography variant='h3'>Career</Typography>
@@ -15,14 +11,15 @@ const CareerSection = ({ width, projects }) => {
         Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua.
         Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua.
       </Typography>
-      <Grid container spacing={5} justify='center' className={classes.projects}>
+      <Grid container spacing={5} justify='center' style={{ marginTop: 75 }}>
         {projects.map(card => (
           <Grid key={card.title} item >
-            <ProjectCard
+            <ProjectSummaryCard
               title={card.title}
               description={card.description}
               technologies={card.technologies}
               projectLink={card.projectLink}
+              elevation={5}
             />
           </Grid>
         ))}
@@ -32,8 +29,5 @@ const CareerSection = ({ width, projects }) => {
 }
 
 CareerSection.propTypes = {
-  projects: PropTypes.array.isRequired,
-  width: PropTypes.string.isRequired
+  projects: PropTypes.array.isRequired
 }
-
-export default withWidth()(CareerSection)