+++ src/reducers/index.js
@@ -1,8 +1,10 @@
 import { combineReducers } from 'redux'
+import { careerPage } from './careerPage'
 import { contactForm } from './contactForm'
 
 const rootReducer = combineReducers({
-  contactForm
+  contactForm,
+  careerPage
 })
 
 export default rootReducer