+++ src/util/technologies.js
@@ -1,0 +1,86 @@
+export const technologies = {
+  AWS_API_GATEWAY: {
+    name: 'Gateway',
+    link: 'https://aws.amazon.com/api-gateway/'
+  },
+  AWS_LAMBDA: {
+    name: 'Lambda',
+    link: 'https://aws.amazon.com/lambda/'
+  },
+  CASSANDRA: {
+    name: 'Cassandra',
+    link: 'http://cassandra.apache.org/'
+  },
+  CSHARP: {
+    name: 'C#',
+    link: 'https://docs.microsoft.com/en-us/dotnet/'
+  },
+  DYNAMODB: {
+    name: 'Dynamo',
+    link: 'https://aws.amazon.com/dynamodb/'
+  },
+  GIT: {
+    name: 'Git',
+    link: 'https://git-scm.com/'
+  },
+  GRADLE: {
+    name: 'Gradle',
+    link: 'https://gradle.com/'
+  },
+  JAVA: {
+    name: 'Java',
+    link: 'https://openjdk.java.net'
+  },
+  JAVASCRIPT: {
+    name: 'JavaScript',
+    link: 'https://www.javascript.com/'
+  },
+  KAFKA: {
+    name: 'Kafka',
+    link: 'https://kafka.apache.org/'
+  },
+  KOTLIN: {
+    name: 'Kotlin',
+    link: 'https://kotlinlang.org/'
+  },
+  KTOR: {
+    name: 'Ktor',
+    link: 'https://ktor.io/'
+  },
+  MICRONAUT: {
+    name: 'Micronaut',
+    link: 'https://micronaut.io/'
+  },
+  MONGO: {
+    name: 'Mongo',
+    link: 'https://www.mongodb.com/'
+  },
+  POSTGRESQL: {
+    name: 'Postgres',
+    link: 'https://www.postgresql.org/'
+  },
+  PROCESSING: {
+    name: 'Processing',
+    link: 'https://processing.org/'
+  },
+  PYTHON: {
+    name: 'Python',
+    link: 'https://www.python.org/'
+  },
+  REACT: {
+    name: 'React',
+    link: 'https://reactjs.org/'
+  },
+  REDUX: {
+    name: 'Redux',
+    link: 'https://redux.js.org/'
+  },
+  SPRING: {
+    name: 'Spring',
+    link: 'https://spring.io/'
+  },
+  TYPESCRIPT: {
+    name: 'TypeScript',
+    link: 'https://www.typescriptlang.org/'
+  }
+}